Changed downloads folder
Now all files exported from the flex monitor are now in the raw folder. This helps for better project structure and for the data processing script. Also, I moved the web scraping script to a folder dedicated for it, leaving the main folder for the later on main code. Changed the name of the folder that contains the functions used though the project to "functions". Added some libraries to the requirements.txt as now I changed the logic of downloading files, first checking if raw folder exists and cleaning all files in it before running the script. I have also improved the logic for saving the log files